training/skeleton/requirements.txt
absl-py==0.11.0
backcall==0.2.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
dgl-cu110==0.5.3
freetype-py==2.2.0
future==0.18.2
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.35.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.4.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.4
networkx==2.5
numpy==1.19.4
oauthlib==3.1.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==3.0.8
pyglet==1.5.14
Pygments==2.7.3
PyOpenGL==3.1.0
pyparsing==2.4.7
pyrender==0.1.43
python-dateutil==2.8.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scipy==1.5.4
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
#torch==1.5.0
#torchvision==0.6.0
traitlets==4.3.3
trimesh==3.9.1
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.4.0
